Refresh needs not to use the Sendable trait
Using sendable would require too many structures to have an additional
template variable while all we really need out of it is the u64
type_tag. This can be easily provided to the refresh function
through arguments. Another advantage is that we'll be able to
avoid the double parsing of the type_tag.