Fixed inconsistent spacing.
The spacing in the functions seem to be inconsistent.  I updated the spacing
between functions to match that found in the rust documentation.

https://doc.rust-lang.org/nightly/book/hello-world.html

According to the rust docs:
There are a number of details that are important here. The first is that it’s indented with four spaces, not tabs. Please configure your editor of choice to insert four spaces with the tab key.