Don't drop proxy nodes.
Dropping the proxy node can prevent responses (in particular connection
infos from close nodes) from reaching the joining node, as the new test
showed. Instead, just let them time out after five minutes.

Since the first connection will be one of the `PARALLELISM` closest
nodes, guaranteeing that responses will reach us, we can send the
`GetCloseGroup` requests to the buckets immediately, instead of waiting
for `GROUP_SIZE` connections.

This change also removes the node join limit, which will be
reimplemented in a more useful way: Instead of the proxy node, the
future close group should, as a response to `ExpectCloseNode`, decide
whether it will accept the new member, and reject it if it is still
dealing with the previous joining or leaving node.