Include the close group in the GetNetworkName response.
Firstly, this saves one round trip, because the client doesn't have to
send a `GetCloseGroup` request to that same group.

Secondly, it limits the frequency of joining nodes to prevent damaging
the integrity of the network: If the group is already in the process of
dealing with another joining or leaving node, the `GetNetworkName`
response won't accumulate and the new node won't join.