Added handling for Backup and Sacrificial chunks to DM.
Affects mainly immutable_data_manager.rs.

The DM now Puts to Backup and Sacrificial locations when receiving a client Put.  It also responds to client requests for these types, and will try and retrieve from alternative locations if required.  This commit also removes a dependency on the LRU crate altogether.  Tests still need to be updated.