Updated DM tests.
Affects mainly immutable_data_manager.rs.

This fixes the DM tests which broke when adding Backup and Sacrificial handling to the DM.  These unit tests could be extended significantly - all code paths are definitely not covered; however the priority for now is the large integration test using mock Crust.  The other changes are mostly due to rustfmt and updated dependencies.