Allow network to report `NetworkFull` to Clients.
Affects mainly immutable_data_manager.rs.

This implements a list of "full" PmidNodes at the IDM, so that they can block Put requests if too many close nodes are full.  No tests are in place for this at the moment.  The remaining changes are due to dependency updates and running rustfmt.