feat/all: handles adding a new contact
This is an initial (temporary) implementation of handling the addition of a new node to our
RoutingTable.  The addition is not synchronised across the group, nor is the handling of the
group split.  Similarly, the new node does nothing to synchronise the incoming lists of contacts
which are effectively the routing tables of all peers in our own group; it just tries to connect
to every contact provided.