MAID-1663: fix/all: fix duplicate handling of messages (#1198)
* fix/all: fix duplicate handling of messages

This is part fix and part workaround for the issue where clients and nodes are handling some
'RoutingMessage's more than once.

* fix/message_filter: change outgoing message filter to use 'RoutingMessage'

This completes the fix for duplicate relaying of messages from a proxy node to a client.  The
'SignedMessageFilter' has been updated (and renamed) to work with 'RoutingMessage's rather than
'SignedMessage's.