fix/signature_accumulator: fix a bug in 'SignatureAccumulator::add_message()'
This fixes a bug which prevented any message from accumulating in the 'SignatureAccumulator'.  It
also adds a couple of basic unit tests for this struct, improves the debug output of 'RoutingTable'
and implements 'Drop' for it so that all RTs will be logged in the case of a test panicking.