MAID-1675: accumulate messages in the first hop (#1213)
* fix/clippy: fix new clippy warning

* chore/build: update dependencies

* fix/utils: update calculate_relocated_name and re-enable and fix test

* feat/node: accumulate messages in the first hop

* fix/routing_table: workaround for routing connection info via the joining node

* fix/routing_table: use the joining node's name for excluding groups

* style/all: address review comments

* fix/messages: address review comments

* fix/messages: address review comments

* fix/routing_table: update group merge prefixes to remain consistently sorted

This resolves an issue where some lists of prefixes related to merging groups were not sorted,
causing the merge test to fail sporadically.

* chore/ci: increase Travis CI timeouts

* fix/node: address further review comments

* fix/xorable: fix bug in 'Xorable::common_prefix()' for arrays

This fixes an issue and adds a regression test for a bug in the 'impl_xorable_for_array' macro.

* fix/signature_accumulator: fix a bug in 'SignatureAccumulator::add_message()'

This fixes a bug which prevented any message from accumulating in the 'SignatureAccumulator'.  It
also adds a couple of basic unit tests for this struct, improves the debug output of 'RoutingTable'
and implements 'Drop' for it so that all RTs will be logged in the case of a test panicking.

* style/clippy: fix clippy warning

* fix/core_test: ignore an intermittently-failing test for now

This causes 'core_tests::more_than_group_size_nodes()' to be ignored.  Also addresses further
review comments.