fix/routing_table: merge if below minimum group size
Merge whenever we drop below `min_group_size`, not only at
`min_group_size - 1`. The original reason for that - not initiating more
than one merge events at the same time - is gone, as the `merging` field
tracks ongoing merge events. But if a merge is actually blocked by that,
we need to make sure that we initiate another one when the next node
leaves.