feat/node: collect only `MIN_GROUP_SIZE` signatures
This is a - probably temporary - workaround for the problems with
sending a group message while merging or splitting. If we require only
the `MIN_GROUP_SIZE` nodes closest to the source address to sign the
message, it works in both cases, and replaces the other workarounds: we
don't need to store group lists in the ack manager anymore, and we can
stop resending messages when we moved out of the sending group.