fix/event: Fix blocking behaviour of next_ev and propagate non-blocking errors. (#1305)
This ensures that calling `next_ev` will continually block and process all
events, necessary for, e.g. `run` in safe_vault.

Additionally, using the mpsc error types allows users of `try_next_ev` to determine
if an error was returned because there were no events available, or because of a permanent failure.