fix/tunnels: fix bug when handling tunnel messages
This fixes a couple of issues in node when handling 'TunnelDirect' messages - one where a tunnel
client would ignore a message intended for it, and another where _any_ 'DirectMessage' being handled
by a tunnel node from a pending tunnel client would cause the node to accept both clients.

Also fixed a todo in tunnels.rs to clean up 'new_clients' when moving a pair to 'clients'.