node/fix: swarm messages to tunnelled peers multiple times
This reduces the chance of a peer not receiving a swarmed message when the rest of its section
believe it has received it.

This commit also increases the lower limit of the network size in the mock crust churn tests and
avoids blocking connections to node 0, since at the start of the tests node 0 is unable to act as a
tunnel node for any other node.