fix/node: send section updates even if RT not valid if accumulated enough requests
This adds an accumulator for 'SectionUpdateRequest' messages to reduce the chance of a single peer
triggering a section-wide response.  Once three such requests are accumulated, the recipients send
a 'SectionUpdate' even if their routing tables show that the invariant is broken.

It also changes how 'SectionUpdates' are handled in that we'll only split neighbouring sections,
but will merge neighbours or our own section.