fix/various: avoid purging RT while handling timeout in mock-crust tests
The purge is avoided in this situation since an enqueued but as yet unhandled 'CrustEvent::LostPeer'
can cause the invariant check to fail.  Checking the invariant at this point is not sound since the
event, once handled, causes the node's internal state to become valid again.

This also fixes two panic-while-panicking issues; one in the 'log_or_panic!' macro and the other in
'Node::drop()'.

It also fixes an inverted condition in 'XorName' causing an error message to be logged incorrectly.