refactor/bootstrappers: update handling of bootstrapping nodes
This adds a new 'PeerState' variant for bootstrapping peers which contains the type of connection
(client or node).  That's used to replace the 'bootstrappers' container in 'Node'.

There are further restrictions placed on the types of messages clients and bootstrappers can send,
and some trivial renaming changes included in this commit.