fix/bootstrap: enforce one client per IP only on bootstrap request (#1506)
* fix/bootstrap: enforce one client per IP only on bootstrap request
Due to the nature of parallel bootstrap attempts by crust it can happen that two clients with same IP successfully trigger a bootstrap accept in the same proxy but one of them decides to choose another proxy. That client will have disconnected from this proxy but the other client, C, could have decided to choose this one. So C gets a bootstrap connect. However if C was the latter of the two for this proxy and if this proxy hasn't yet realised that the first client has disconnected before it got the bootstrap accept for C, it is going to reject C on basis of one-client-per-IP rule. So C will get a disconnect event after it has received a bootstrap connect.
Moving this check to Bootstrap request should prevent this as only the ones who are interested in this proxy will be sending the bootstrap request and hence get promoted to a client.

* refactor/review: address review points