Create routine model following the flow.md file (only the destination)
Contains basic tests.
We can consider property tests to validate it.

The graph is replicated in the 300 first lines.
However a number of helpers are needed in the next 300.
The rest is test code.

Note: This does not really follow best practices rust, it tries to
provide an as concise code and as close as possible to the flow.