Update Flow to handle Node getting/loosing elder status + top flow
Routine ending make the sub routine be killed.
This is useful if an elder node is relocated, then it should not try
to process node acceptance.

Also, since it is no longer an elder, none of its current processing
is relevant anymore. Maybe it should pass along messages for some times?
Do we want to add a timed buffer between loosing elder status and being
relocated, and elder passing messages.