Update Flow to avoid coupling changing elder with other flows.
There were difficulties in having mutliple elder flow in parrallel.
This flow avoid this and gives complete control to when elder are changed.

Here a simple consensused timer is used, but any other shared
state condition can be used.

When relocating, the node is marked as relocating, so if it is an elder,
it is not relocated yet, it will wait until once the check elder as updated
its status to non elder, then on the next relocation trigger it will
start relocating the node (a bit like it would happen if the first time
it was refused).