fix up RelocatingState to avoid accepting to many new nodes
We want to not accept relocating node if either resource proof is not
complete or we are still waiting to relocate the previous relocated
node.

To make things clean, and ensure the relocating routine as enough
information to send the expected age with Expectcandidate,
RelocatingState now come with an enum {AgeIncrease, Hop, BackOnline).

This also ensure that they can be treated differently:
 - Only relocate one of our adults at a time, but disregard other nodes
   we are relocating.
 - Only relocate one of our relocated node to a new hop, but disregard
   other nodes we are relocating.
 - Only relocate back online node when all the others are relocated
   to ensure they do not interfere with existing running nodes.