fix/routing: add NetworkEvent::PurgeCandidate for candidate fail
We want to ensure eventually that the candidate is a state shared by
all peers in the section. To do that all modification must be the
result of reaching consensus on an event.

We use a new consensused event PurgeCandidate to decide together
whether a candidate failed or succeeded.

Note:
-Add new state: AcceptedWaitingSectionInfo.
 This avoid us accepting a candidate while the previous candidate has
 not caught up yet (i.e before the SectionInfo that contains it is
 consensused).
-Because this is a consensused timer, we use Candidate expired_once
 to ensure we only vote for it once. Other peers will do the same.

Test:
Run local soak test.