fix/routing: add Add/RemoveElder event, make Online/Offline Opaque.
If a candidate fail resource proof we do not want to add it to our
members at all. Because Parsec Add make Parsec consider the peer
as valid voter, we needed to Remove it afterward if it failed:
 => Do not use Parsec Add to consensus on whether to accept or reject
    a candidate: Use new Online/Offline opaque events.

This will also be essential for Node Ageing as most new peers will
not be voters in Parsec (i.e they will be Adults and not elder).

Test:
New unit tests
Soak tests.