fix/routing: keep proxy connection for JoiningNode after split
Fix soak test failure:
messages_during_churn:
Some([1699640792, 566487471, 1587518420, 2229705965])
(Adapted from Viv patch suggestion).

The proxy for the joining node get notify of the Section with the
JoiningNode before dispatching Node approaval.
At the same time, it is also part of a split section, so it decide
to disconnect the joining node that it thinks is a full node.

This bring back a solution that was implemented before:
Do not remove a node that we knew as joining, put it back to be a
joining node and only disconnect it when time out is reached.

Note:
Add tracing for messages dropped on disconnection that would have
more immediatly show the issue, and also occur in limited cases.

Test:
Use seed, and patch to ensure behaviour change only apply to this one
joining node: Verify seed now pass.