fix/routing: vote online when resource proof complete
candidate_timer_token was used before we moved to Parsec to ensure
the RPC from different elders would be sent at the same time.

This is no longer needed with Parsec: Send it when resource proof is
complete.
It was also acting as a rate limiter, but it was an unintended
consequence.