feat/rmd: Disable chain verification in hop
The code removed here ensured that the message from our neighbour was
trusted, hence establishing a chain of trust as the message travelled
through the Network.

All this logic will be overhauled with the Secure Message Delivery
implementation. For now, it needs to be removed so the tests can pass
with the other aspects of Reliable Message Delivery.