remove UserMessage chunking
Replace `MessageContent::UserMessagePart` with `MessageContent::UserMessage` which contains the whole message. Remove all the logic around splitting the message into parts and reassembling it back.

This is possible to do because quic-p2p internally re-prioritizes the messages based on size, so big messages no longer block the small ones and so it's no longer necessary to chunk them.