feat/mock: pretend quic-p2p exposes `SentUserMessage`
Change the mock to expose what we wish the implementation would expose.
We will update the actual implementation once the mock feature is
convenient enough for the features we need in routing.

In this case, we would like for quic-p2p to inform its users once a
message cannot result in an `UnsentUserMessage` anymore.
This will remove the need for routing to keep its own timeouts in a way
or another (probably as a `LruTimeCache`).