perf/quic_p2p: avoid many serialization on send/use Rc
Spandan higlighted that Bytes is a reference counted type.
This clarified that we were being inefficient in memory and CPU when
sending 3 messages.

quic_p2p take in a Bytes type that is a reference counted type:
 - Refactor to do cheap operation multiple time (Rc clone) instead of
   expansive operation (serialize + allocate memory for message 3 times).
 - Refactor mock_serialize to follow the same pattern.

Ensure all messages go through the same resend process and hide the
token implementation detail.