feat/routing: disable merge in tests - panic on merge
Merge functionality is partially broken and will be more as
development continue ignoring it. New functionality will break down
on Merge.

We may want to completely strip the merge functionality out since it
is now untested and will rapidely bit-rot requiring a rewrite if we
want to re-add it.

-Ensure that no test proceed to trigger merge: panic if merge occur.
-Update churn tests to not trigger merge, but still cover churn and
 prefix change (split).

Test:
Run soak tests.