feat/routing: ensure part of our state is shared our_infos/our_history
This place the building block of sharing state with only part of
SharedState. Follow up will add their_keys, and neighbour_infos as
weel as their_knowledge.

Disable prunning our_infos, this is required until their_knowledge
is solid and shared state as we require updates to our_infos to be
shared by all peers in exactly the same way.

Test:
Run soak tests.