feat/routing: remove proofs from NeighbourSigs and ensure shared state
Since we are not using ProvingSections, we do not need to Neighbour
SectionInfo proofs. Remove un-needed code. Re-voting for proving section
was ensuring test passed, so only removing it fail the tests.

To ensure the test continue to pass, at the same time ensure shared
state is shared through genesis block. neighbour_infos should be in
SharedState, will move it in later change.

As a result, we can also remove re-voting for neighbour_infos sections.

Test:
clippy & test pass