fix/routing: update their_knowledge for correct prefix
Their knowledge is a map of their prefix to our version that they know.
The AckMessage need to contain the prefix, and the consensused message
need to consensus that prefix with the version.

We incorrectly used our own section that we consensused and then update
our prefix and version.

Test:
Update check_section_info_ack so it fails without this fix.
Run clippy+ soak test