fix/routing: fix split their_knowldge updating their_keys on new event
SectionInfo on split will only output a single section, but both will
be consensused. As a result, it will not be consensused again when
receiving messages.

Using the same mechanism to consensus ourselves and the neighbour we
see has this problem. Additionally, we will need different set of info
whether we get a message or NeighbourInfo.

Test:
Updated check_section_info_ack
Run soak tests.