feat/routing: ensure security_metadata contain the needed info
We need to be able to create SectionKeyInfo from security_metadata.
Also, we need to trust that SectionKeyInfo: To do so, we follow the
same pattern established with SectionInfo, that contain prefix and
version number, and these 3 fields form the item that need to be signed
by the previous key to prove descendance.

Replace BlsPublicKey by SectionKeyInfo in the required place.
Currently implement SectionKeyInfo using the underlying SectionInfo that
exists in BlsPublicKey. When we switch to real BLS, SectionKeyInfo will
have 3 fields: real BlsPublicKey, version, prefix.

The sender_prefix is now bundled with the proof.
We now use the security_metadata instead of src_section to generate
TheirKeyInfo. This will be necessary to remove the src_section.

Test:
Clippy + soak test.