feat/routing: do not implicitly trust unsigned messages
All routing messages should be signed as we then use the content to
make decision within routing base on the accuracy on the information.

Unsigned messages need to be handled separately as we cannot trust its
content. Currently they are not used, but for Vault to send RMD messages
this will probably be unsigned messages (though they could also be
single source signed like we do for relocate requests here).

Test:
Clippy + soak test