refactor/chain: extract ChainAccumulator and add tests
chain_accumulator & completed_events work together.
We will need to update them to handle signing BLS key, so extract them
and add unit test.

Note:
In this shape it become clear that an invariant of the struct is that
no event in `completed_events` can be in `chain_accumulator`, so
we should be able to later safely remove the check in reset_accumulator.

Test:
New unit test.
clippy + soak tests.