feat/parsec: keep track of size of requests/responses
Keep track of size and number of requests/responses so we can
prune/reset the parsec gossip graph before it becomes too bloated.

Note that we use serialised_size to accumulate the size. This may turn
out to have a performance impact so we might need to address this
properly in the future. Maybe to have parsec tell us the
request/response size.