feat/chain: Signature share place holder for NetworkEvent::SectionInfo
We want to be able to sign the new BLS Key that we will consensus
eventually with the previous BLS Key. We may also want to sign the
SectionInfo with the new BLS key.

Eventually we will have something like:
NetworkEvent::SectionInfo(SectionInfo, SectionKeyInfo, Signature)

The current code use None for this place holder in production and
chain_accumulator unit test validate the working with a real value.
Eventually we will not use an option.

Note:
Considered an alternative design where NetworkEvent would contain
the current NetworkEvent enum and the option for the signature.

Test:
Run clippy + Unit test + soak tests.