fix/elder: Not vote for already consensused neighbour SectionInfo
We use to accumulate signatures for neighbours SectionInfo.
With Secure Message Delivery, that was no longer necessary.
But we did not remove these additional unecessary votes.

We want to add signature field for SectionInfo votes, and it would
not be possible to sign with the same signature as all the other votes,
so we remove this code now.

Test:
Clippy + soak tests.