refactor/chain: clarify SectionInfo accumulate and clean dead code
SectionInfo have been accumulating for a while only if we have quorum
with the current SectionInfo. As a result we do not need to check
that again in add_section_info (was used in other placed before).

`is_valid_neighbour_info` is also unused, so we can remvoe the
associated dead code.

Test:
Run soak test + clippy