Send gossips also as Adult, after accumulating AddElder of self
This prevents a potential problem when a new node might timeout while waiting for the `SectionInfo` which contains itself. This is because the existing nodes might send it only gossips of the post-split version which the new node cannot handle. By allowing it to send gossips itself, the response to those gossips would have the correct version allowing to Adult to make progress.