refactor/node: refactor Node API to be used by Vaults
This change makes `StateMachine::step` non-blocking and consequently allows `Node`
to process routing events in a non-blocking fashion.

As a result of this refactoring, we can now process arbitrary events within the
scope of the Routing event loop, allowing us to confine the user code (e.g. Vaults)
to the same thread as Routing.

This is achieved by using the `crossbeam_channel::Select` abstraction which allows to
register multiple event receivers and listen on all of them at the same time.