fix/routing: process unbroken chain of NeighbourInfo
With seed Some([2634247339, 1651835408, 2311746235, 352137822]) failing
on multiple_joining_nodes.

We now process the neighbour in the correct order, but we fail to get
quorum on some Neighbour.

Test:
Verify with logs that we now have neighbour processed in correct order,
however, we are blocked for some neighbours to get the needed quorum.