fix/routing: do not ignore NeighbourInfo if we already got a latter one
Fix seed Some([2634247339, 1651835408, 2311746235, 352137822]) failing
on multiple_joining_nodes,

The problem is that we got NeighbourInfo out of order.
As a result we vote for a later NeighbourInfo, and then when the initial
once come, we discard it as no longer relevant. But it is relevant since
we expect an unbroken chain of NeighbourInfo.

Also it remove code that is uncessary since Parsec will already filter
things we already voted for.

Test:
Seed now pass.
Clippy pass + soak test