test/routing: update multiple_joining_nodes to avoid recursive split
The multiple_joining_nodes test added random nodes. As a result with
seed Some([802906621, 3015886335, 2758576582, 1229525089]), the
situation occured where:

2 split occur (1 -> 11 -> 111) before the Neighbour 0 had time to update
its neighbour info. When 0 update its section info it sends it to prefix(1),
and that message is dropped because 111 neighbours do not cover the full
1 prefix (anti spam prevention since a message could be spammed to
everyone in the network if Destination prefix is not narrow enough.

Removing that spam prevention in chain::targets "fix" the issue.
However, the correct way to fix that issue is to ensure this kind of
recursive split does not occur in our tests as it should not happen in
the real network either.

Test:
clippy + soak test