fix/states: remove timeout for Adult to transition to Elder
This failed with Seed Some([2536718493, 1108160782, 926471325, 1679426090])
on check_section_info_ack with features=mock_base.

The problem is that this did not leave enough iteration for Parsec to
reach consensus on both AddElder and the new SectionInfo.

This timeout does also not make sense in the context of node ageing
since most Adult will stay adults and not transition to Elder.

Test:
Clippy + Soak test + mock_base with seed passing.