fix/chain: handle new neighbours that are the result of a split
Fix seed Some([2410421335, 3238787189, 896801798, 350710283])
in multiple_joining_nodes where a new NeighbourInfo was received before
we caught up to the split. We were ignoring it, but we need to vote for
it so it can be re-voted for after the split and allow to reach
consensus.

Test:
Verify seed now work. Clippy + soak test