refactor/parsec: simplify elder test using test function to vote
Add a new vote_for_as function to the ParsecMap that can only be used
with mock_parsec, and that allow using a single ParsecMap to write
Elder unit test.

This helps as well with handling split where a new genesis Payload
need to be voted for.

Test:
clippy + unit test pass (initially developped with earlier commit as
parent to verify unit test that were not ignored still pass).