fix/chain: allow send message to old member after split
Fix node_pause_and_resume seed:
Some([2_996_986_743, 943_531_375, 1_508_150_456, 2_954_527_262])

NodeApproval is sent to 2 nodes that have not yet seen Online.
 => They cannot send the message so send to everyone.
Everyone but the pause node and new node is in sibling, by the time
they try to send the message, split_self has executed and they cannot
send the message.

Ensure we can still relay message to our ex-members after a split.

Note: This may be better handled in the future by avoiding accumulation
for NodeApproval similarly to how we do it for Relocate. For example,
the approved node could be the target of signature accumulation and do
the accumulation itself.

Test:
Verify seed pass.