fix/mock: quic_p2p always return success or failure
An assumption of the system is that quic_p2p will always return a
success or failed status for a message. However, this was not always
happening.

As a result, a message sent to a dropped node was not necessarily
getting a failed status, and combined with another node being relocated,
would result in a message being dropped/stuck un transmitted.

Additionally we did not clear the pending messages on Packet::Disconnect.
So we could get pending messages left un-reported as failed. so call
clear_pending_messages there as well.

Test:
Verify seed is fixed when calling clear_pending_messages only from the
point the problem was found to not break the seed: aggressive_churn:
Some([4185393102, 1125699391, 3155230365, 1759798662])
soak test + clippy.