Fix flaw in the relocation request flow that could cause no response
The flaw was that the decision whether to accept or deny a relocate request was evaluated in a way which does not guarantee consensus. Because of this, it could happen that a relocate request would be neither accepted, nor denied - that is, the requesting section would receive no response.

The fix is to first vote through parsec and only then evaluate the decision.