feat/routing: sign SectionKeyInfo and added to NetworkEvent::SectionInfo
When trying to remove bls_emu, one of the problem was the missing
information and signatures for the SectionKeyInfo.

The simplest way to address this is to add the SectionKeyInfo to the
payload, as that avoid having to store information locally and allow
clear flow of execution:

DKG
-> vote SectionInfo(ElderInfo + SectionKeyInfo with DKG result + sig)
-> consensus: Update our info and key. New key signed by old one.

Note:
- This does not address the need of SectionKeyInfo to be passed the new
  RealBlsPublicKey when real BLS is used.
- our_section_next_bls_keys_sig now does not use the new_dkg optional.
  It should not have used it, as all nodes should sign with the previous
  key they hold the new key, whether they are part of that new section
  or not.

Test:
Clippy + tests.