feat/routing: use RealSectionKeyInfo in NetworkEvent::SectionInfo
To pass around the RealBlsPublicKey, we need to pass along
RealSectionKeyInfo in NetworkEvent::SectionInfo.

To achieve that we need to make the RealBlsPublicKey accessible in
a number of places. Doing that here initially will help the bls_emu
removal as RealSectionKeyInfo will already be constructible in all the
places it needs to be.

Test:
Clippy + soak test