Make handling of denied relocate requests more robust
Previous it could happen that a node would receive `RelocateRequestDenied` message before it even accumulated the corresponding `RelocateRequest` event. It would consider the message invalid and not retry the request. If this happened to enough nodes, the relocate request process could get stalled.

The fix is to always vote for the retried request and only check whether it was valid after it accumulates.