fix/chain: derive implementation of PartialEq...
Some of these were problematic to derive before real BLS was in place
since combined signature could be different while both valid. Not an
issue with BLS.

Also fix buggy implementation of PartialEq that required both to be
valid for equality to hold, and so failed when both were invalid.

Test:
Soak + clippy