Make sure messages are trusted by all nodes during ongoing split
Consider this situation:

A section `P` splits into `P0` and `P1`. Say `P0` already updates some other section `R` about their knowledge. Then `R` sends a message to `P0`. `P1` might not have fully caught itself to the split so it still thinks its prefix is `P` and so it considers itself in authority of that message. But if the message is proven only for `P0`, then `P1` would not trust it.

The fix is to construct the proof so that it is still trusted by both `P0` and `P1` until `P1` also realizes the split and updates `R` about their knowledge.